Inital commit of fenbuild based off some previous scripts i made
Fenbuild is a commandline tool for easily creating, setting up and using
pbuilder. The idea behind the tool is to direct a user through creating
a pbuilder environment so they can focus on packaging.

I've tried to model the command line interface similar to git in the sense
that some state is saved as to which environment you current have selected
which makes it easier when building for a particular environment or
multiple environments with the same package.