Extra checking and consistent exit statuses
I wanted the current build environment to work better so I've added
checks to see if the directory exists and if it's in the available builds
file before allowing you to change build environments.

I added the feature to be able to set the current build with the second
argument as opposed to having to use the interactive select screen each
time.

Added better defined exit statuses for scripting use.